requirements.txt
fastapi[all]
requests
httpx
openai
pypdf
python-docx
boto3
pyjwt
dotenv==0.9.9
uvicorn
python-dotenv
pydantic
redis
motor
flower
eventlet
PyJWT==2.10.1
pymongo==4.12.0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
gunicorn
celery
python-flutterwave
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
limits
bcrypt
APScheduler==3.11.0
authlib
celery-aio-pool
